+++ packages/acs-client/src/acs/interfaces.ts
@@ -54,8 +54,8 @@ export interface Database {
 
 export interface AuthZSubject {
   id: string; // entity ('user', 'service', etc) ID
-  role_associations: RoleAssociation[];
-  hierarchical_scopes?: HierarchicalScope[];
+  // role_associations: RoleAssociation[];
+  // hierarchical_scopes?: HierarchicalScope[];
 }
 
 export interface HierarchicalScope {
@@ -144,7 +144,7 @@ export interface AuthZResponse extends Response {
   obligation: string;
 }
 
-export interface IAuthZ extends AuthZ<AuthZSubject | UnauthenticatedData, AuthZContext, Resource[], AuthZAction> {
+export interface IAuthZ extends AuthZ<Subject | UnauthenticatedData, AuthZContext, Resource[], AuthZAction> {
   whatIsAllowed: (request: Request<AuthZWhatIsAllowedTarget | NoAuthWhatIsAllowedTarget, AuthZContext>,
     ctx: ACSClientContext, useCache: boolean) => Promise<PolicySetRQResponse>;
 }