+++ packages/oidc-id-srv-integration/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/restorecommerce/libs/compare/@restorecommerce/oidc-id-srv-integration@1.0.1...@restorecommerce/oidc-id-srv-integration@1.0.2) (2022-10-04)
+
+**Note:** Version bump only for package @restorecommerce/oidc-id-srv-integration
+
+
+
+
+
 ## [1.0.1](https://github.com/restorecommerce/libs/compare/@restorecommerce/oidc-id-srv-integration@1.0.0...@restorecommerce/oidc-id-srv-integration@1.0.1) (2022-08-25)
 
 **Note:** Version bump only for package @restorecommerce/oidc-id-srv-integration