+++ packages/service-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://github.com/restorecommerce/libs/compare/@restorecommerce/service-config@1.0.0...@restorecommerce/service-config@1.0.1) (2022-10-04)
+
+**Note:** Version bump only for package @restorecommerce/service-config
+
+
+
+
+
 # [1.0.0](https://github.com/restorecommerce/libs/compare/@restorecommerce/service-config@0.4.28...@restorecommerce/service-config@1.0.0) (2022-08-25)